Remove local virtual environment setup from GitHub workflow
The sections of the GitHub workflow related to setting up a local virtual environment have been removed. This change came about as these instructions were redundant in the system. The instructions to create and position the virtual environment within the project are now handled within the poetry.toml file configuration.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>