Update general unit test in SchemArrow module
The property-based general unit test for the SchemArrow module has been modified. A copy of the pandas DataFrame (df) is now created before it's passed to SchemArrow. Verification has been added to ensure the original df remains unmodified after passing through SchemArrow and every dtype in the resulting DataFrame is of type 'pyarrow'.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>