Rename parameter 'default_to_source' in SchemArrow
This command changes the parameter name 'default_to_source' to 'default_target_type' in the SchemArrow class for increased clarity and accuracy. This makes it easier to comprehend the purpose of the parameter, which is to specify the default data type to use when no mapping is found for a specific source type.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>