Added converter instance to module imports
Added an instance of the PandasArrowConverter class, "convert_to_pyarrow", to the imports available in the pandas_pyarrow module. This eliminates the need to create a new instance each time the conversion is needed. It simplifies the process and makes the code clearer, especially for new users. This change has been reflected into the README documentation as well.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>