Remove Codecov token from CI workflow
The CODECOV_TOKEN has been removed from the ci.yml file. By omitting the token, we rely on the Codecov GitHub Action's default behavior, which automatically provides the proper token for repositories that are public. However, no change in functionality is expected since the Codecov Action does work without the token for public repositories on GitHub.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>