Update version of various Python packages
Updated the versions of google-cloud-bigquery, hypothesis, pyasn1, pyasn1-modules, and ruff libraries in the poetry.lock file. The changes also include updating python-versions compatibility and some dependency files hashes. This commit is aimed at keeping the software up-to-date with the latest features and security fixes provided with the new versions.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>