Update paths for triggering code checks on pull requests (#172)
* Update paths for triggering code checks on pull requests

Added specific paths to the pull_request trigger to include changes in `pandas_pyarrow`, `tests`, and the workflow configuration file itself. This ensures that code checks run for relevant modifications, improving workflow efficiency and coverage.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

* update dep.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

* Fix halffloat conversion in reverse_converter.py

Update the dtype check for halffloat to use .startswith rather than an exact match. This ensures consistent handling of various halffloat type representations in the conversion process.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

* Update paths for triggering code checks on pull requests

Added specific paths to the pull_request trigger to include changes in `pandas_pyarrow`, `tests`, and the workflow configuration file itself. This ensures that code checks run for relevant modifications, improving workflow efficiency and coverage.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

* Fix halffloat conversion in reverse_converter.py

Update the dtype check for halffloat to use .startswith rather than an exact match. This ensures consistent handling of various halffloat type representations in the conversion process.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

* Set up Codecov for main branch on push

Added configuration to trigger Codecov report uploads on push events to the main branch. This ensures that code coverage data is kept up-to-date with the latest changes in the main branch.

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>

---------

Signed-off-by: DanielAvdar <66269169+DanielAvdar@users.noreply.github.com>